requirements.txt
crewai
langchain
langchain-google-genai
google-generativeai
python-dotenv
fastapi
uvicorn
pydantic